test: 비즈니스 예외처리를 Service단에서 처리
