+++ src/test/java/roomescape/repository/reservationtime/ReservationTimeH2RepositoryTest.java
@@ -39,13 +39,6 @@ class ReservationTimeH2RepositoryTest {
     }
 
     @Test
-    @DisplayName("중복된 시간을 저장하려고 하면 예외가 발생한다.")
-    void saveDuplicatedGetTime() {
-        assertThatThrownBy(() -> reservationTimeH2Repository.save(RESERVATION_TIME_1))
-                .isInstanceOf(UserException.class);
-    }
-
-    @Test
     @DisplayName("id에 맞는 ReservationTime을 제거한다.")
     void delete() {
         reservationTimeH2Repository.delete(RESERVATION_TIME_2.getId());
