+++ src/test/java/roomescape/repository/theme/ThemeH2RepositoryTest.java
@@ -39,15 +39,6 @@ class ThemeH2RepositoryTest {
     }
 
     @Test
-    @DisplayName("중복된 테마를 저장하려고 하면 예외가 발생한다.")
-    void saveDuplicatedTime() {
-        Theme theme = new Theme(null, THEME_1.getName(), "레벨2 설명", "레벨2 썸네일");
-
-        assertThatThrownBy(() -> themeH2Repository.save(theme))
-                .isInstanceOf(UserException.class);
-    }
-
-    @Test
     @DisplayName("id에 맞는 Theme을 제거한다.")
     void delete() {
         themeH2Repository.delete(THEME_2.getId());
